feat: extrating components part 3
