+++ lib/live_view_phoenix_web/live/volunteers_live.ex
@@ -2,17 +2,14 @@ defmodule LiveViewPhoenixWeb.VolunteersLive do
   use LiveViewPhoenixWeb, :live_view
 
   alias LiveViewPhoenix.Volunteers
-  alias LiveViewPhoenix.Volunteers.Volunteer
 
   def mount(_params, _session, socket) do
     if connected?(socket), do: Volunteers.subscribe()
     volunteers = Volunteers.list_volunteers()
-    changeset = Volunteers.change_volunteer(%Volunteer{})
 
     socket =
       assign(socket,
         volunteers: volunteers,
-        changeset: changeset,
         recent_activity: nil
       )
 
@@ -24,37 +21,6 @@ defmodule LiveViewPhoenixWeb.VolunteersLive do
 
   #   """
   # end
-
-  def handle_event("save", %{"volunteer" => params}, socket) do
-    case Volunteers.create_volunteer(params) do
-      {:ok, _volunteer} ->
-        # socket =
-        #   update(
-        #     socket,
-        #     :volunteers,
-        #     fn volunteers -> [volunteer | volunteers] end
-        #   )
-
-        changeset = Volunteers.change_volunteer(%Volunteer{})
-
-        socket = assign(socket, changeset: changeset)
-
-        {:noreply, socket}
-
-      {:error, %Ecto.Changeset{} = changeset} ->
-        socket = assign(socket, changeset: changeset)
-        {:noreply, socket}
-    end
-  end
-
-  def handle_event("validate", %{"volunteer" => params}, socket) do
-    changeset =
-      %Volunteer{}
-      |> Volunteers.change_volunteer(params)
-      |> Map.put(:action, :insert)
-
-    {:noreply, assign(socket, changeset: changeset)}
-  end
 
   def handle_info({:volunteer_created, volunteer}, socket) do
     socket =
