+++ lib/live_view_phoenix_web/live/volunteers_live.html.leex
@@ -1,16 +1,8 @@
 <%= live_component @socket, LiveViewPhoenixWeb.HeaderComponent, title: "Volunteer Check-In", subtitle: @recent_activity %>
 <div id="checkin">
-<%= f = form_for @changeset, "#", phx_submit: "save", phx_change: "validate", class: "form-checkin"%>
-<div class="field">
-  <%= text_input f, :name, placeholder: "Name", autocomplete: "off" %>
-  <%=error_tag f, :name %>
-</div>
-<div class="field">
-  <%= telephone_input f, :phone, placeholder: "Phone", autocomplete: "off" %>
-  <%=error_tag f, :phone %>
-</div>
-<%= submit "Check In",phx_disable_with: "Saving..." %>
-</form>
+  <div>
+    <%= live_component @socket, LiveViewPhoenixWeb.FormVolunteerComponent, id: :new %>
+  </div>
   <div >
     <%= for volunteer <- @volunteers do %>
       <%= live_component @socket, LiveViewPhoenixWeb.VolunnterComponent, volunteer: volunteer, id: volunteer.id %>
